refactor: QRException 수정
